requirements.txt
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
